Element 3: the 159 new species, records and graphics
Every species on the pick-list now exists in the tree with real data and real
art. pl_personal, evo and wotbl go from 508 members to 667, pl_pokegra from
2964 to 3918, the icons from 547 to 706, footprints from 497 to 656 and the
sprite offsets from 1976 to 2612, all of which the species-ID scheme predicted
from one insertion into generated/species.txt. SPECIES_EGG is 653 now, so
MAX_SPECIES is 654 and NATIONAL_DEX_COUNT is 652.

Three sources, each authoritative for different fields, with the split and the
evidence for it in docs/oxide/donor-tables.md. Base stats, types, both
abilities and the hidden ability come from New Pokedex.xlsx, which is Ian's
design record and matches all 159 by dex_pos and name. Catch rate, gender
ratio, hatch cycles, friendship, exp rate, egg groups, body colour, safari flee
rate, base experience, icon palette, learnsets, evolutions, names, categories,
heights, weights, sprites, palettes, icons and cries come from the donor.

Two engine changes the data forced, both small:

- Base experience was a byte and Generation 7 values run to 340 for these
  species, 39 of which are over 255. It moved to the two bytes of padding the
  compiler was already leaving at the end of the record, so the record is still
  48 bytes and nothing between catch rate and the abilities shifted.
- MAX_LEARNSET_ENTRIES was 20 and three of the new species have more; it is 34
  now, matching the donor's own fixed slot count so it will not need raising
  again for this source. move_reminder_data.c's buffer was a hand-written 22
  and now derives from it, which is what it was all along.

What is deliberately short, all of it recorded in the tracker: TM, tutor and
egg-move lists are empty, because Hardlove keeps TM compatibility somewhere
this project has not found and the TM list itself is a Phase 5 pass, so filling
them now would be work thrown away. Dex entry text is a placeholder, because
the donor has none for these species. Body shape and the dex size-comparison
scales are defaults. Footprints are blank, with "has": false. Level-up
learnsets are filtered to Platinum's own 467 moves, which keeps 80 per cent of
the entries and leaves no species empty; element 4 refills them.

Verified: make rom succeeds; verify_narcs reports 0 disagreeing, 20 intended
Fairy retypes and 159 new species across all three per-species archives, every
native record still matching the base ROM field for field; the base ROM
importer still wants to change nothing; the three bulk dry-runs are still
clean. verify_narcs learned two things to get there, that the three archives
are built from one registry whose indices after the natives have all moved, and
where base experience went.

Not done here: the seven natives that gain an evolution into a new species
(Primeape, Scyther, Eevee, Gyarados, Lopunny, Wooper, Luvdisc). That is a
separate pass because it edits records the base ROM carried over.

+++ tools/dataproc/src/speciesproc.c
@@ -512,7 +512,7 @@ static SpeciesData proc_personal(datafile_t *df) {
             enum_u8(".egg_groups[1]", enum EggGroup),
         },
 
-        .baseExpReward  = u8(".base_exp_reward"),
+        .baseExpReward  = u16(".base_exp_reward"),
         .baseFriendship = u8(".base_friendship"),
         .bodyColor      = (u8)(enum_u8(".body_color", enum PokemonColor) & maxbit(7)),
         .catchRate      = u8(".catch_rate"),
